% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr_fxs.R
\name{build_fits_df}
\alias{build_fits_df}
\title{Organize estimates and likelihood from probability model fits}
\usage{
build_fits_df(x)
}
\arguments{
\item{x}{list of probability distribution fits. Probably a list of outputs from fitdistrplus::fit_dist()}
}
\value{

}
\description{
Organize estimates and likelihood from probability model fits
}
\examples{

}
